feat(autoAgg): Make components reasonable
